WEB-139: add support for CPU and model description
also add 404 for invalid model or cpu; add tests
