+++ archives/templates/computers/model.html
@@ -5,6 +5,9 @@
 <p class="nav">{% include "computers/navbar.html" %}</p>
 
 <section class="archives">
+  {% if intro %}
+  <div class="intro">{{ intro|safe }}</div>
+  {% endif %}
   <div id="yearindex">
       {% for ad in ads %}
           <div>
